requirements.txt
streamlit==1.28.0
langchain>=0.1.20
langchain-community>=0.0.38
langchain-openai>=0.1.8
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
openai>=1.40.0,<2.0.0
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==3.17.4
numpy>=1.25.0
pandas>=2.1.3

